factored commands out to a separate project.  some of the relative directory paths used to locate the library resources in cli-library-manager were no longer valid, so the system now fetches the location by invoking the module, to get the directory from the source of truth.
