+++ accept/bin/libraryResources.js
@@ -1,0 +1,5 @@
+/**
+ * Outputs the location of the library test resources provided by the library manager package.
+ */
+var libraryTestResources = require('particle-commands').libraryTestResources;
+console.log(libraryTestResources());
