+++ accept/features/library_install_deps.feature
@@ -7,7 +7,7 @@ Feature: installation of libraries with dependencies
   Scenario: ensure libraries are published
     Given The particle library "test-library-transitive-1" is removed
 
-    Given I copy the directory "../../../node_modules/particle-cli-library-manager/resources/libraries/publish/transitive" to "."
+    Given I copy the library resource "publish/transitive" to "."
     And a directory named "transitive/trans1" should exist
     And I cd to "transitive/trans1"
     And I run particle "library publish"
