+++ accept/features/library_install_nodeps.feature
@@ -3,7 +3,7 @@ Feature: library install
 
   Scenario: help does mention vendor
     When I run particle "library install --help"
-    Then the output should not contain "--vendored"
+    Then the output should contain "--vendored"
 
   Scenario: help does not mention adapter
     When I run particle "library install --help"
@@ -17,20 +17,9 @@ Feature: library install
     And the file "lib/neopixel/library.properties" should exist
     And the file "lib/neopixel/src/neopixel.cpp" should exist
     And the file "lib/neopixel/src/neopixel.h" should exist
-    And the file "lib/neopixel/src/neopixel/neopixel.h" should not exist
-    And the exit status should be 0
-
-  Scenario: as a user I can install neopixel as a vendored library in an extended project with adapter headers
-    Given an empty file named "project.properties"
-    And an empty file named "src/hello.cpp"
-    When I run particle "library install --vendored --adapter neopixel"
-    Then the directory "lib" should exist
-    And the file "lib/neopixel/library.properties" should exist
-    And the file "lib/neopixel/src/neopixel.cpp" should exist
-    And the file "lib/neopixel/src/neopixel.h" should exist
     And the file "lib/neopixel/src/neopixel/neopixel.h" should exist
     And the exit status should be 0
-
+    
   Scenario: as a user I can attempt to install a non-existent library as a vendored library in an extended project
     Given an empty file named "project.properties"
     And an empty file named "src/hello.cpp"
