+++ package.json
@@ -68,7 +68,6 @@
     "moment": "^2.9.0",
     "node-wifiscanner2": "^1.2.0",
     "particle-api-js": "^6.0.6",
-    "particle-cli-library-manager": "0.0.4-develop.5",
     "request": "^2.46.0",
     "semver": "^5.1.0",
     "serialport": "^3.1.1",
@@ -98,6 +97,7 @@
     "istanbul": "^0.3.22",
     "mocha": "^3.0.2",
     "mock-fs": "^3.11.0",
+    "particle-cli-library-manager": "0.0.4-develop.5",
     "proxyquire": "^1.6.0",
     "rimraf-promise": "^2.0.0",
     "should": "^7.0.2",
@@ -136,7 +136,7 @@
     "coveralls": "istanbul cover ./node_modules/mocha/bin/_mocha --report lcovonly -- -R spec --compilers js:babel-register test/ && cat ./coverage/lcov.info | ./node_modules/coveralls/bin/coveralls.js && rm -rf ./coverage",
     "doctoc": "doctoc --title '## Table of Contents' README.md",
     "lint": "eslint -f unix src/**/*.js",
-    "test": "mocha test/ test/oldcmd/ test/app test/cli test/cmd test/integration --compilers js:babel-register",
+    "test": "mocha test/ test/oldcmd/ test/app test/cli test/integration --compilers js:babel-register",
     "compile": "babel src -d dist",
     "prepublish": "npm run compile",
     "update-firmware-binaries": "node scripts/update-firmware-binaries.js"
