--- src/cmd/commands.js
@@ -1,11 +1,0 @@
-
-// export all commands. Ideally the commands would be in a separate package, but for now they are bundled with the CLI.
-
-export * from './library';
-export * from './library_delete';
-export * from './library_init';
-export * from './library_install';
-export * from './library_migrate';
-export * from './library_publish';
-export * from './library_search';
-export * from './project_properties';
