--- src/cmd/readme.md
@@ -1,1 +1,0 @@
-UI-neutral command framework and command implementations.
