--- test/cmd/library_install.spec.js
@@ -1,17 +1,0 @@
-
-import {expect, sinon} from '../test-setup';
-import {LibraryInstallCommand, LibraryInstallCommandSite} from "../../src/cmd/library_install";
-
-
-
-
-
-describe('library install', () => {
-
-	it("installs all dependencies when no library is present", () => {
-		const sut = new LibraryInstallCommand();
-		const site = new LibraryInstallCommandSite();
-		
-	});
-
-});
