--- test/fixtures/projects/simple/expectedProperties.json
@@ -1,5 +1,0 @@
-{
-  "name": "My project",
-  "sentence": "Short description of the project",
-  "dependencies.internetbutton": "1.0.0"
-}
