--- test/fixtures/projects/simple/project.properties
@@ -1,3 +1,0 @@
-name=My project
-sentence=Short description of the project
-dependencies.internetbutton=1.0.0
